音频/音乐/音乐-机器学习/requirements.txt
absl-py==0.13.0
astunparse==1.6.3
flatbuffers==1.12
google-pasta==0.2.0
grpcio==1.34.1
h5py==3.1.0
keras==2.6.0
keras-preprocessing==1.1.2
pandas==1.4.0
numpy==1.19.5
opt-einsum==3.3.0
protobuf==3.17.3
tensorboard==2.6.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow-estimator==2.6.0
tensorflow-gpu==2.6.0
termcolor==1.1.0
typing-extensions==3.7.4.3
wheel==0.36.2
wrapt==1.12.1
scipy==1.6.2
music21